Move ShaderPresetAddon to ADDON namespace
